+++ src/lib/sh.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
 import path from 'node:path'
-import { get_logger } from './logger'
+import { get_logger } from './log'
 import { get_config } from './config'
 import os from 'node:os'
 import child_process from 'node:child_process'